requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
black==21.7b0
click==8.0.1
Django==3.2.5
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumbers==8.12.27
Pillow==8.3.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.0
